refactor: extract task execution context builder
